+++ node/src/main/kotlin/net/corda/node/internal/security/RPCSecurityManagerWithLocalShellUser.kt
@@ -1,0 +1,34 @@
+package net.corda.node.internal.security
+
+import net.corda.node.services.config.shell.localShellUser
+import net.corda.nodeapi.internal.config.User
+import org.apache.shiro.mgt.DefaultSecurityManager
+import org.apache.shiro.subject.SimplePrincipalCollection
+import javax.security.auth.login.FailedLoginException
+
+/**
+ * Wrapper for [RPCSecurityManager] which creates in-memory [AuthorizingSubject] for [User].
+ * Can be used to add on a specific [User] on top of the principals provided by the [RPCSecurityManager] realm.
+ * By the default grants "ALL" permissions to "shell/shell" user/password.
+ */
+class RPCSecurityManagerWithLocalShellUser(private val delegate: RPCSecurityManager, private val user: User = localShellUser()) : RPCSecurityManager by delegate {
+
+    private val realmId = user.username + "Realm"
+    private val shellAuthorizingSubject = ShiroAuthorizingSubject(subjectId = SimplePrincipalCollection(user.username, id.value),
+            manager = DefaultSecurityManager(InMemoryRealm(listOf(user), realmId)))
+
+    @Throws(FailedLoginException::class)
+    override fun authenticate(principal: String, password: Password): AuthorizingSubject =
+            if (user.username == principal && user.password == password.valueAsString) {
+                shellAuthorizingSubject
+            } else {
+                delegate.authenticate(principal, password)
+            }
+
+    override fun buildSubject(principal: String): AuthorizingSubject =
+            if (user.username == principal) {
+                shellAuthorizingSubject
+            } else {
+                delegate.buildSubject(principal)
+            }
+}
